@@ -2,11 +2,11 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Bullseye" Version="3.0.0-beta.3" />
+    <PackageReference Include="Bullseye" />
     <PackageReference Include="McMaster.Extensions.CommandLineUtils" />
     <PackageReference Include="SimpleExec" />
   </ItemGroup>
