@@ -4,7 +4,6 @@ using Microsoft.AspNetCore.Hosting;
 using Serilog;
 using Serilog.Events;
 using Serilog.Sinks.SystemConsole.Themes;
-using Microsoft.Extensions.Logging;
 
 namespace SampleApi
 {
@@ -30,11 +29,7 @@ namespace SampleApi
 
             return WebHost.CreateDefaultBuilder(args)
                     .UseStartup<Startup>()
-                    .ConfigureLogging(builder =>
-                    {
-                        builder.ClearProviders();
-                        builder.AddSerilog();
-                    })
+                    .UseSerilog()
                     .Build();
         }
     }
