@@ -1,4 +1,4 @@
-﻿﻿$(function () {
+﻿window.addEventListener("load", function () {
     var a = document.querySelector("a.PostLogoutRedirectUri");
     if (a) {
         window.location = a.href;
