@@ -3,6 +3,7 @@
 
 
 using Host.Configuration;
+using IdentityModel;
 using IdentityServer4;
 using IdentityServer4.Quickstart.UI;
 using Microsoft.AspNetCore.Builder;
@@ -10,8 +11,10 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Logging;
 using Microsoft.IdentityModel.Tokens;
+using Serilog;
 using System.Linq;
 using System.Security.Claims;
+using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
 
 namespace Host
@@ -29,10 +32,17 @@ namespace Host
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllersWithViews()
-                .AddNewtonsoftJson();
+            services.AddControllersWithViews();
 
+            // configures IIS out-of-proc settings (see https://github.com/aspnet/AspNetCore/issues/14882)
             services.Configure<IISOptions>(iis =>
+            {
+                iis.AuthenticationDisplayName = "Windows";
+                iis.AutomaticAuthentication = false;
+            });
+
+            // configures IIS in-proc settings
+            services.Configure<IISServerOptions>(iis =>
             {
                 iis.AuthenticationDisplayName = "Windows";
                 iis.AutomaticAuthentication = false;
@@ -52,16 +62,13 @@ namespace Host
                 //.AddInMemoryClients(_config.GetSection("Clients"))
                 .AddInMemoryIdentityResources(Resources.GetIdentityResources())
                 .AddInMemoryApiResources(Resources.GetApiResources())
-                .AddDeveloperSigningCredential()
+                .AddSigningCredential()
                 .AddExtensionGrantValidator<Extensions.ExtensionGrantValidator>()
                 .AddExtensionGrantValidator<Extensions.NoSubjectExtensionGrantValidator>()
                 .AddJwtBearerClientAuthentication()
                 .AddAppAuthRedirectUriValidator()
                 .AddTestUsers(TestUsers.Users)
                 .AddMutualTlsSecretValidators();
-
-            //var key = CryptoHelper.CreateECDsaSecurityKey();
-            //builder.AddSigningCredential(key, SecurityAlgorithms.EcdsaSha256);
 
             services.AddExternalIdentityProviders();
             services.AddLocalApiAuthentication(principal =>
@@ -91,11 +98,12 @@ namespace Host
 
         public void Configure(IApplicationBuilder app)
         {
+            app.UseSerilogRequestLogging();
+
             app.UseDeveloperExceptionPage();
             app.UseStaticFiles();
 
             app.UseRouting();
-            app.UseMiddleware<Logging.RequestLoggerMiddleware>();
             app.UseIdentityServer();
 
             app.UseAuthorization();
@@ -104,6 +112,33 @@ namespace Host
             {
                 endpoints.MapDefaultControllerRoute();
             });
+        }
+    }
+
+    public static class BuilderExtensions
+    {
+        public static IIdentityServerBuilder AddSigningCredential(this IIdentityServerBuilder builder)
+        {
+            // create random RS256 key
+            //return builder.AddDeveloperSigningCredential();
+
+            // use an RSA-based certificate with RS256
+            //var cert = new X509Certificate2("./keys/identityserver.test.rsa.p12", "changeit");
+            //return builder.AddSigningCredential(cert, "RS256");
+
+            // ...or PS256
+            //return builder.AddSigningCredential(cert, "PS256");
+
+            // or manually extract ECDSA key from certificate (directly using the certificate is not support by Microsoft right now)
+            var cert = new X509Certificate2("./keys/identityserver.test.ecdsa.p12", "changeit");
+            var key = new ECDsaSecurityKey(cert.GetECDsaPrivateKey())
+            {
+                KeyId = CryptoRandom.CreateUniqueId(16)
+            };
+
+            return builder.AddSigningCredential(
+                key,
+                IdentityServerConstants.ECDsaSigningAlgorithm.ES256);
         }
     }
 
