@@ -1,5 +1,4 @@
 ﻿using IdentityModel;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Tokens;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
@@ -88,17 +87,13 @@ namespace IdentityServer4.Configuration
         {
             var signingAlgorithmBits = int.Parse(signingAlgorithm.Substring(signingAlgorithm.Length - 3));
 
-            switch (signingAlgorithmBits)
+            return signingAlgorithmBits switch
             {
-                case 256:
-                    return SHA256.Create();
-                case 384:
-                    return SHA384.Create();
-                case 512:
-                    return SHA512.Create();
-                default:
-                    throw new InvalidOperationException($"Invalid signing algorithm: {signingAlgorithm}");
-            }
+                256 => SHA256.Create(),
+                384 => SHA384.Create(),
+                512 => SHA512.Create(),
+                _ => throw new InvalidOperationException($"Invalid signing algorithm: {signingAlgorithm}"),
+            };
         }
 
         /// <summary>
@@ -106,17 +101,13 @@ namespace IdentityServer4.Configuration
         /// </summary>
         internal static ECCurve GetCurveFromCrvValue(string crv)
         {
-            switch (crv)
+            return crv switch
             {
-                case JsonWebKeyECTypes.P256:
-                    return ECCurve.NamedCurves.nistP256;
-                case JsonWebKeyECTypes.P384:
-                    return ECCurve.NamedCurves.nistP384;
-                case JsonWebKeyECTypes.P521:
-                    return ECCurve.NamedCurves.nistP521;
-                default:
-                    throw new InvalidOperationException($"Unsupported curve type of {crv}");
-            }
+                JsonWebKeyECTypes.P256 => ECCurve.NamedCurves.nistP256,
+                JsonWebKeyECTypes.P384 => ECCurve.NamedCurves.nistP384,
+                JsonWebKeyECTypes.P521 => ECCurve.NamedCurves.nistP521,
+                _ => throw new InvalidOperationException($"Unsupported curve type of {crv}"),
+            };
         }
 
         /// <summary>
