@@ -7,6 +7,8 @@ using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using IdentityServer4.Extensions;
+using Microsoft.Extensions.DependencyInjection;
+using IdentityServer4.Configuration;
 
 namespace IdentityServer4
 {
@@ -33,6 +35,8 @@ namespace IdentityServer4
             IEnumerable<Claim> additionalClaims = null)
         {
             var claims = new HashSet<Claim>(new ClaimComparer());
+            var context = tools.ContextAccessor.HttpContext;
+            var options = context.RequestServices.GetRequiredService<IdentityServerOptions>();
             
             if (additionalClaims != null)
             {
@@ -52,7 +56,11 @@ namespace IdentityServer4
                 }
             }
 
-            claims.Add(new Claim(JwtClaimTypes.Audience, string.Format(IdentityServerConstants.AccessTokenAudience, tools.ContextAccessor.HttpContext.GetIdentityServerIssuerUri().EnsureTrailingSlash())));
+            if (options.EmitLegacyResourceAudienceClaim)
+            {
+                claims.Add(new Claim(JwtClaimTypes.Audience, string.Format(IdentityServerConstants.AccessTokenAudience, tools.ContextAccessor.HttpContext.GetIdentityServerIssuerUri().EnsureTrailingSlash())));
+            }
+            
             if (!audiences.IsNullOrEmpty())
             {
                 foreach (var audience in audiences)
