@@ -168,7 +168,9 @@ namespace IdentityServer4.IntegrationTests.Clients
                     AllowedScopes =
                     {
                         "api1", "api2"
-                    }
+                    },
+
+                    AllowOfflineAccess = true
                 },
                 new Client
                 {
